Examples - common Rust example code in shared local crate
