+++ examples/rust/ex1-volume/Cargo.toml
@@ -9,3 +9,4 @@ edition = "2018"
 [dependencies]
 structopt = { version = "0.3", default-features = false }
 libceed = { path = "../../../rust" }
+mesh = { path = "../mesh" }
